chore: merged in changes from 3.0 branch
