+++ src/immutable/float/construction.rs
@@ -1,5 +1,5 @@
 use crate::float::kdtree::Axis;
-use crate::immutable_float::kdtree::ImmutableKdTree;
+use crate::immutable::float::kdtree::ImmutableKdTree;
 use crate::types::Content;
 use std::ops::Rem;
 use tracing::{event, Level};
