+++ src/lib.rs
@@ -82,8 +82,7 @@ mod custom_serde;
 pub mod distance_metric;
 pub mod fixed;
 pub mod float;
-//pub mod float_sss;
-pub mod immutable_float;
+pub mod immutable;
 mod mirror_select_nth_unstable_by;
 pub mod nearest_neighbour;
 #[doc(hidden)]
